Create lib/dom.ts
